task 0

+++ 0x0B-malloc_free/0-create_array.c
@@ -19,7 +19,7 @@ char *create_array(unsigned int size, char c)
 		return (NULL);
 
 	char_arr = malloc(sizeof(char) * size);
-	for (i = 0; i < size ; i++)
+	for (i = 0; i < (int)size ; i++)
 	{
 		char_arr[i] = c ;
 	}
